ml/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.14.0
transformers==4.35.2
pytest==7.4.3
python-dotenv==1.0.0
pydantic==2.5.2
joblib==1.3.2
nltk==3.8.1
spacy==3.7.2
pytorch==2.1.1
matplotlib==3.8.2
seaborn==0.13.0
scipy==1.11.4
optuna==3.4.0
mlflow==2.8.1
category_encoders==2.6.2
feature_engine==1.6.1
imbalanced-learn==0.11.0